separate interfaces and implementations
