+++ src/Detector/Implementations/CMakeLists.txt
@@ -1,0 +1,6 @@
+target_sources(faces
+        PRIVATE
+        OcvDefaultDnnDetector.cpp
+        PUBLIC
+        OcvDefaultDnnDetector.h
+        )
