+++ src/Recognizer/CMakeLists.txt
@@ -5,3 +5,4 @@ target_sources(faces
         )
 
 add_subdirectory(Descriptors)
+add_subdirectory(Implementations)
