+++ src/Recognizer/Implementations/Descriptors/DlibResnetDescriptor.cpp
